added feat node dist from line and removed node option from project.jsx to increase readability of code
