+++ client/src/utils/constants.js
@@ -1,3 +1,4 @@
 export const threshold = 10;
 export const fixedWidth = 200;
 export const fixedHeight = 200;
+export const spacing = 400;
